+++ backend/src/main/java/com/example/backend/dto/response/permission/PermissionInfoResponse.java
@@ -2,21 +2,21 @@ package com.example.backend.dto.response.permission;
 
 public class PermissionInfoResponse {
 
-    private Long id;
+    private Integer id;
     private String name;
 
     public PermissionInfoResponse() {}
 
-    public PermissionInfoResponse(Long id, String name) {
+    public PermissionInfoResponse(Integer id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
