+++ backend/src/main/java/com/example/backend/repository/ChapterRepository.java
@@ -11,12 +11,12 @@ import java.util.List;
 @Repository
 public interface ChapterRepository extends JpaRepository<Chapter,Integer> {
     // Lấy danh sách chapter theo course, sắp xếp theo orderIndex
-    List<Chapter> findByCourse_IdOrderByOrderIndexAsc(Long course_id);
+    List<Chapter> findByCourse_IdOrderByOrderIndexAsc(Integer course_id);
 
     // Cập nhật orderIndex khi drag & drop
     @Modifying
     @Query("UPDATE Chapter c SET c.orderIndex = :orderIndex WHERE c.id = :id AND c.course.id = :courseId")
-    int updateOrderIndex(@Param("id") Integer id, @Param("courseId") Long courseId, @Param("orderIndex") Integer orderIndex);
+    int updateOrderIndex(@Param("id") Integer id, @Param("courseId") Integer courseId, @Param("orderIndex") Integer orderIndex);
 
 }
